App and Environment: shuffle & extend `ContentSizeCategory`
This moves the source file used for the `ContentSizeCategory` extensible
enumeration.  It adds some missing interfaces as well.
